pyproject.toml
[project]
name = "email-triage-env"
version = "0.2.0"
description = "Email Triage Environment for Scaler x Meta OpenEnv Hackathon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "watchfiles>=0.21.0",      # for reload in dev (optional but useful)
]

[project.scripts]
# This is the critical entry point the deployment checker looks for
server = "server.app:create_app"

[tool.uv]
dev-dependencies = []

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.uv.sources]
# If you need any local or git dependencies later

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
]

[project.urls]
Homepage = "https://github.com/snarkeesbanu-saleem/email-triage-env"